Add exit button functionality.
Added save and exit functionality to all pages. Save and exit
function is in the utilities function and the button was added in
action.html and popup.html. On click functionality added to
buttons in overlayScreen.js and prewalkthrough.js. Also fixed
tool tip functionality in action.js Also updated alert for missing
team name and missing scenario name in prewalkthrough.js.