Add dynamic pronouns for personas.
Added dynamic pronouns for tool tip pop-ups. This also includes
updating some of the verbs in the informational paragraphs. For this,
the updates are made to is/are and other verbs that would add an s to
the end of the word when the pronoun is she or him. This is done by
wrapping words in a span tag with a specific class name and then
changing the text of the span in the javascript file.