Hide save and exit button when on exit screen.
Added functionality to remove the save and exit button when use is
on the exit screen. This is because it would break the tool to let the
use click it repeatedly and may have been confusing for a user since
they already clicked it. Button reappears if the user returns to
session without exiting.