Add none of the above checkbox functionality.
The none of the above check box data was not being saved to memory thus
it was not populating when the user checked it, moved on, and then used
the back button to return. The check boxes were not connected to the
data saving process properly. Updated html element id's and some
javascript to match other checkbox functionality.