Add option to download old format.
Added option on exit screen to download the old format of the GM
session csv file. It has the same format as the 'download file again'
button and implements the old functions for creating the csv. Also
added bool variable to the functions for downloading to determine
the name of the file.