GenderMag Persona/Styles Documentation
Purpose
The documentation that I wrote aims to solve the problem of understanding what the HCI concepts of personas and information processing/learning styles are. For new users of GenderMag, they might not understand what the purpose of the project is because they might not be familiar with these concepts and how they relate to usability engineering and design. So, I wrote a document that explains these concepts and gives examples, as well as provides links to the GenderMag website for even more clarification. The goal of this documentation was to provide very simple and easy-to-read explanations of these concepts, instead of diving deep into these concepts and confusing the users.

Files  - 1 new file created

Issue tested and verified - None
Comments - None